requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
selenium==4.27.1